Move helpMenuLinks to separate file
